# backend/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package database

import (
	"database/sql"
	"database/sql/driver"
	"fmt"
	"time"

	"github.com/google/uuid"
	"github.com/sqlc-dev/pqtype"
)

type SurchargeType string

const (
	SurchargeTypePercentage SurchargeType = "percentage"
	SurchargeTypeExact      SurchargeType = "exact"
)

func (e *SurchargeType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = SurchargeType(s)
	case string:
		*e = SurchargeType(s)
	default:
		return fmt.Errorf("unsupported scan type for SurchargeType: %T", src)
	}
	return nil
}

type NullSurchargeType struct {
	SurchargeType SurchargeType
	Valid         bool // Valid is true if SurchargeType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullSurchargeType) Scan(value interface{}) error {
	if value == nil {
		ns.SurchargeType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.SurchargeType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullSurchargeType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.SurchargeType), nil
}

type IntegratorConfig struct {
	ID                 uuid.UUID
	ClientID           sql.NullString
	ProviderID         sql.NullInt32
	Name               sql.NullString
	DisplayName        sql.NullString
	IntegratorName     sql.NullString
	SpID               sql.NullString
	PlazaIDMap         pqtype.NullRawMessage
	Extra              pqtype.NullRawMessage
	Url                sql.NullString
	TaxRate            sql.NullString
	Surcharge          sql.NullString
	SurchangeType      NullSurchargeType
	InsecureSkipVerify sql.NullBool
	CreatedAt          time.Time
	UpdatedAt          time.Time
}

type IntegratorTransaction struct {
	ID                    uuid.UUID
	BusinessTransactionID uuid.UUID
	Lpn                   sql.NullString
	IntegratorID          uuid.NullUUID
	Status                sql.NullString
	Amount                sql.NullString
	Error                 sql.NullString
	Extra                 pqtype.NullRawMessage
	TaxData               pqtype.NullRawMessage
	CreatedAt             time.Time
	UpdatedAt             time.Time
}

type Log struct {
	ID        uuid.UUID
	Level     sql.NullString
	Message   sql.NullString
	Fields    pqtype.NullRawMessage
	CreatedAt time.Time
}

type OaTransaction struct {
	ID                    uuid.UUID
	Businesstransactionid string
	Lpn                   sql.NullString
	Customerid            sql.NullString
	Jobid                 sql.NullString
	Facility              sql.NullString
	Device                sql.NullString
	Extra                 pqtype.NullRawMessage
	EntryLane             sql.NullString
	ExitLane              sql.NullString
	CreatedAt             time.Time
	UpdatedAt             time.Time
}

type SnbConfig struct {
	ID       uuid.UUID
	Name     sql.NullString
	Username sql.NullString
	Password sql.NullString
	Endpoint sql.NullString
	Facility []string
	Device   []string
}

type User struct {
	ID          uuid.UUID
	Name        sql.NullString
	Username    sql.NullString
	Password    sql.NullString
	Permissions []string
}